Records Metric
We  are introducing a new metric which will be the number of records
that have been processed through the proxy. This was previously the
requests counter metric but this is misleading, so now the requests
counter metric is a count of the number of requests and the new records
metric is a count of the number of the records that the proxy had
processed. So this should be a lot more intuitive.